requirements.txt
asgiref==3.5.2
autopep8==1.7.0
Django==4.1.2
django-js-asset==2.0.0
django-modeltranslation==0.18.5
django-mptt==0.14.0
django-tinymce==3.5.0
Pillow==9.2.0
pycodestyle==2.9.1
python-dotenv==0.21.0
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
